- Replace url strings with contants in URLS enum. - Move useSignIn.ts into hooks/services
