+++ src/constants/urls.ts
@@ -1,7 +1,7 @@
 enum URLS {
-    SERVER_BASE = "http://locahost:8002",
-    SEVER_USER_SIGN_IN = `${SERVER_BASE}/sign-in`,
-    SEVER_USER_SIGN_UP = `${SERVER_BASE}/sign-up`,
+    SERVER_BASE = "http://localhost:8002",
+    SEVER_USER_SIGN_IN = `${SERVER_BASE}/user/sign-in`,
+    SEVER_USER_SIGN_UP = `${SERVER_BASE}/user/sign-up`,
 }
 
 export default URLS;
