+++ src/hooks/services/useSignIn.ts
@@ -1,13 +1,14 @@
 import axios, { AxiosError } from "axios";
-import { IUser, userSignedIn } from "../store/userStore/userSlice";
+import { IUser, userSignedIn } from "../../store/userStore/userSlice";
 import { useDispatch } from "react-redux";
+import URLS from "../../constants/urls";
 
 const useSignIn = () => {
     const dispatch = useDispatch();
 
     const signInUser = async (email: string, password: string) => {
         try {
-            const { data } = await axios.post("http://localhost:8002/user/sign-in", {
+            const { data } = await axios.post(URLS.SEVER_USER_SIGN_IN, {
                 email: email,
                 password: password,
             });
@@ -16,7 +17,7 @@ const useSignIn = () => {
             dispatch(userSignedIn(user));
         } catch (error) {
             if (error instanceof AxiosError) {
-                console.log(error);
+                console.log(error.response?.data);
             }
         }
     };
